refactor formatting utils
